requirements.txt
# AutoPinner Pro Dependencies
# Generated by setup_dependencies.py

PyQt5>=5.15.0
requests>=2.31.0
cryptography>=42.0.0
PyJWT>=2.8.0
psutil>=5.9.0
beautifulsoup4>=4.12.0
selenium>=4.18.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiohttp>=3.9.0
pillow>=8.0.0
python-dateutil>=2.8.0
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
